Notebook - SQUID magnetisation vs. temperature
This program extracts ".dat" data (or similar) from Quantum Design SQUID outputs, for the extraction of magnetisation vs. temperature curves. It performs the derivative of the magnetisation vs. the temperature for the finding of the material's Curie temperature. It also includes an example of smoothing using the scipy.signal module.